server/requirements.txt
requests
flask
python-dotenv
flask_sqlalchemy
pymysql